+++ src/OpenTheWindows.Cli/CommandLineBuilder.cs
@@ -25,6 +25,7 @@ internal static class CommandLineBuilder
         root.Subcommands.Add(ScanCommand.Create(services));
         root.Subcommands.Add(HealthCommand.Create(services.Health));
         root.Subcommands.Add(ApplyCommand.Create(services));
+        root.Subcommands.Add(RemediateCommand.Create(services));
         root.Subcommands.Add(RevertCommand.Create(services));
         root.Subcommands.Add(HistoryCommand.Create(services));
         root.Subcommands.Add(UpdatesCommand.Create(services));